s/o/detect: More precise check of octave-dev-pkg tools usage
Ensure that only packages which include the octave-pkg.mk file from
the octave-pkg-dev package will be selected.  This restriction is
necessary because the generated autopkgtest command relies on the
infrastructure provided by the octave-pkg-dev package, namely the use
of a CDBS-based debian/rules that accepts the 'check-pkg' target.

Also, update the comment to make this point clear.