Octave: the generated test depends now only on octave-autopkgtest
The generated autopkgtest command for Octave packages corresponds now
to a direct call of the check-pkg script, which will be stripped from
the octave-pkg-dev package and put into a very thin package that will
be called octave-autopkgtest.  This will allow testing the Octave
add-on pacakges without installing the whole building infrastructure.