Don't test non-default python3 for packages with python3-dev build-dep.
By default we do an import test for all versions of python3 returned by
py3versions -r.  But if a package build-depends on python3-dev, indicating
that it builds binary extensions for the current python3, this will fail
whenever there is more than one supported version of python3 since the
extension in the package will by design never have been built for those
additional versions.  Detect this case and only test against py3versions -d.

If the package build-depends on python3-all-dev, or if it doesn't
build-depend on python3-dev at all, then the previous behavior is fine.